pyproject.toml
[tool.poetry]
name = "django-template"
version = "0.1.0"
description = "A template repository for Django projects"
authors = ["Marcelo Lino <mdslino@gmail.com>"]
readme = "README.md"
packages = [{include = "django_template"}]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.7"
gunicorn = "^21.2.0"
python-decouple = "^3.8"
gevent = "^23.9.1"
django-health-check = "^3.17.0"
dj-database-url = "^2.1.0"
psycopg = {extras = ["binary", "pool"], version = "^3.1.12"}
django-structlog = "^6.0.0"


[tool.poetry.group.test.dependencies]
pytest-dotenv = "^0.5.2"
pytest-django = "^4.7.0"


[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
isort = "^5.12.0"
mypy = "^1.7.0"
flake8 = "^6.1.0"
flask-debugtoolbar = "^0.13.1"

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 79

[tool.black]
line-length = 79
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # The following are specific to Black, you probably don't want those.
  | blib2to3
  | tests/data
  | profiling
)/
'''
# We use preview style for formatting Black itself. If you
# want stable formatting across releases, you should keep
# this off.
preview = true

[tool.pytest.ini_options]
env_files = [
    ".test.env"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
